requirements.txt
iniconfig==2.0.0
packaging==24.0
paho-mqtt==2.0.0
pluggy==1.4.0
phonenumbers == 8.13.47
paho-mqtt == 2.1.0
python-dotenv==1.0.1
django-debug-toolbar == 4.4.6
django-environ == 0.11.2
# Django and Core Libraries
Django==4.2.4         # Specify the version you are using
django-phonenumber-field == 8.0.0

# Database Drivers
psycopg2-binary==2.9.6  # For PostgreSQL
# mysqlclient==2.1.1     # Uncomment for MySQL

# Linting and Code Quality
pylint==2.17.5          # For linting (used by SonarCloud)

# Testing Tools
pytest==8.1.1          # For running tests
pytest-django==4.5.2    # Pytest plugin for Django
pytest-cov==4.0.0       # For test coverage

# Security Packages (optional but recommended)
django-environ==0.10.0  # For managing environment variables
django-extensions==3.2  # Useful development extensions
coverage==7.6.1
# Other dependencies used by your project